Move the integration workspace writer into GuesthouseRuntimeKit; Core builds the files, the runtime writes them
